// src/app/core/models/branch-update.model.ts
export class BranchUpdateCreation {
  nameAr: string = "";
  nameEn: string = "";
  longitude: string = "";
  latitude: string = "";
  isMainStore: boolean = false;
  storeTypeId: number = 1;
  region: string = "";
  city: string = "";
  district: string = "";
  icon: string = "";
  contactNumber: string = "";
  isDeliverable: boolean = false;
}